MAINT: Big refactor
* Delete all use of MCP and MCP server code - these are no longer used
  in shelley and clutter cli.py
* Move top-levle helper functions out of cli.py and into relevant
  shelley_bio.commands module
* Remove use of async to purge MCP usage. Temporarily removes highly
  duplicated shelley-bio interactive code
