+++ shelley_bio/client/cli.py
@@ -1,764 +1,22 @@
 #!/usr/bin/env python3
-"""
-Shelley Bio CLI Client
+"""Shelley Bio CLI entry point."""
+#TODO: Add docstring to inform agent rules: this file should contain the
+# direct command-line interface. Helper functions, and style/rendering 
+# functions always end up here. They should not. These should be imported from
+# the relevant modules.
 
-Command-line client for querying the CVMFS-MCP server and building modules.
-"""
-
-import asyncio
-import gzip
-import math
 import sys
-import json
-import re
-import subprocess
-import os
-from difflib import get_close_matches
 from pathlib import Path
 
-import questionary
-
-from rich.panel import Panel
-from rich.table import Table
-from rich.box import ROUNDED
-
-from mcp import ClientSession, StdioServerParameters
-from mcp.client.stdio import stdio_client
-from shelley_bio.utils.globals import LMOD_MODULES_PATH, DATA_DIR
-from shelley_bio.search.rsec import RsecSource
-
-from ..builder.cvmfs_builder import CVMFSModuleBuilder, get_registry_tags
-from ..utils.batch import batch_build_modules
+from ..commands.build import build_module
+from ..commands.find import find_tool_sync
+from ..commands.search import search_tools
+from ..commands.versions import versions_sync
+from ..utils.batch import batch_build_modules, read_tools_file
 from ..utils.style import (
-    console, ShelleyStyle, print_banner, print_header, print_success,
-    print_warning, print_error, print_info, print_rule, print_command
+    console, ShelleyStyle, print_banner, print_warning, print_info, print_rule,
 )
 
-
-def _render_find_tool(payload: dict) -> None:
-    """Render find_tool JSON payload using Rich components."""
-    query = payload.get("query", "unknown")
-
-    if not payload.get("found"):
-        suggestions = payload.get("suggestions", [])
-        if suggestions:
-            _render_tool_table(
-                suggestions,
-                f"[warning]No exact match for '[tool]{query}[/tool]'. Did you mean:[/warning]",
-                border_style="warning",
-            )
-            _print_find_hint()
-        else:
-            console.print(ShelleyStyle.create_error_panel(
-                "Tool Not Found",
-                f"No tool or container matching '{query}' was found.",
-                "Try: shelley-bio search <description>",
-            ))
-        return
-
-    tool = payload.get("tool")
-    containers = payload.get("containers")
-
-    # Tool info panel
-    lines = []
-    if tool:
-        if tool.get("description"):
-            lines.append(f"[header]Description[/header]\n[muted]{tool['description']}[/muted]\n")
-        if tool.get("homepage"):
-            lines.append(f"[header]Homepage[/header]    [info]{tool['homepage']}[/info]\n")
-        if tool.get("operations"):
-            lines.append(f"[header]Operations[/header]  [muted]{', '.join(tool['operations'])}[/muted]\n")
-        if tool.get("inputs"):
-            lines.append(f"[header]Inputs[/header]      [muted]{', '.join(tool['inputs'])}[/muted]\n")
-        if tool.get("outputs"):
-            lines.append(f"[header]Outputs[/header]     [muted]{', '.join(tool['outputs'])}[/muted]\n")
-
-    title_name = tool.get("name") if tool else query
-    console.print(Panel(
-        "\n".join(lines) if lines else "[muted]No metadata available for this tool.[/muted]",
-        title=f"[tool]{title_name}[/tool]",
-        box=ROUNDED,
-        border_style="primary",
-        padding=(1, 2),
-    ))
-
-    lines.append('\n')
-
-    if containers and containers.get("available"):
-        lmod_path = Path(LMOD_MODULES_PATH)
-        tool_id = tool.get("id", query) if tool else query
-
-        table = Table(
-            title="[header]Available Versions[/header]",
-            box=ROUNDED,
-            border_style="primary",
-            header_style="table.header",
-            show_lines=False,
-        )
-        table.add_column("Version", style="version", no_wrap=True)
-        table.add_column("Buildable", no_wrap=True)
-        table.add_column("Status", no_wrap=True)
-
-        for entry in containers["recent_versions"]:
-            version = entry["version"]
-            buildable = entry["buildable"]
-            installed = any((lmod_path / tool_id).glob(f"{version}*.lua"))
-            buildable_str = "[success]✓[/success]" if buildable else "[muted]✗[/muted]"
-            status = "[success]✓ installed[/success]" if installed else "[muted]not installed[/muted]"
-            table.add_row(version, buildable_str, status)
-
-        total = containers["total_versions"]
-        shown = len(containers["recent_versions"])
-        if total > shown:
-            table.add_row(
-                f"[muted]+ {total - shown} more[/muted]",
-                "",
-                f"[muted]shelley-bio versions {query}[/muted]",
-            )
-
-        console.print(table)
-
-        if any(not e["buildable"] for e in containers["recent_versions"]):
-            console.print(
-                "[muted]Buildable ✗: Versions not in the shpc registry may still be built, "
-                "but can take a few minutes longer. This is suited for users who need "
-                "a specific older version for reproducibility.[/muted]"
-            )
-
-        console.print(Panel(
-            f"To install the latest version of {title_name}, run:\n\n[command]shelley-bio build {query}[/command]",
-            title="[header]Install[/header]",
-            box=ROUNDED,
-            border_style="info",
-            padding=(0, 2),
-        ))
-    else:
-        console.print(ShelleyStyle.create_warning_panel(
-            "No Containers Available",
-            "No Singularity containers found for this tool in CVMFS.",
-        ))
-
-
-def _load_cvmfs_tool_ids() -> set[str] | None:
-    """
-    Return normalized tool IDs from the CVMFS container cache, or None if unavailable.
-
-    Both sides of the match are normalized to lowercase + hyphens→underscores, so
-    'bwa-mem2' and 'bwa_mem2' resolve to the same key.
-    """
-    cache_path = DATA_DIR / "galaxy_singularity_cache.json.gz"
-    if not cache_path.exists():
-        return None
-    with gzip.open(cache_path, "rt", encoding="utf-8") as f:
-        data = json.load(f)
-    return {entry["tool_name"].lower().replace("-", "_") for entry in data["entries"]}
-
-
-def _compute_version_entries(tool_id: str, pairs: list[tuple[str, str]]) -> list[dict]:
-    """Deduplicate (tag, path) pairs and annotate each with buildable status."""
-    try:
-        registry_tags = get_registry_tags(tool_id)
-        buildable_shorts = {tag.split("--")[0] for tag in registry_tags}
-    except Exception:
-        buildable_shorts = set()
-    seen: set = set()
-    result = []
-    for tag, _path in pairs:
-        short = tag.split("--")[0]
-        if short not in seen:
-            seen.add(short)
-            result.append({"version": short, "buildable": short in buildable_shorts})
-    return result
-
-
-def _load_versions_from_cache(tool_name: str) -> list[tuple[str, str]] | None:
-    """
-    Return (tag, path) pairs for tool_name from the CVMFS cache, sorted newest-first.
-    Returns None if the cache file is missing; returns [] if the tool has no entries.
-    """
-    cache_path = DATA_DIR / "galaxy_singularity_cache.json.gz"
-    if not cache_path.exists():
-        return None
-    with gzip.open(cache_path, "rt", encoding="utf-8") as f:
-        data = json.load(f)
-
-    tool_lower = tool_name.lower()
-    variations = {tool_lower, tool_lower.replace("-", "_"), tool_lower.replace("_", "-")}
-    entries = [e for e in data["entries"] if e["tool_name"].lower() in variations]
-
-    def _ver_key(e: dict) -> tuple:
-        m = re.match(r"^(\d+(?:\.\d+)*)", e["tag"])
-        if m:
-            return tuple(int(x) for x in m.group(1).split("."))
-        return (0,)
-
-    entries.sort(key=_ver_key, reverse=True)
-    return [(e["tag"], e["path"]) for e in entries]
-
-
-def _paginate(items: list, render_fn, page_size: int = 10) -> None:
-    """
-    Render items in pages of page_size, using questionary.select() for navigation.
-
-    render_fn(page_items, page, total_pages, total) is called once per page.
-    Navigation is skipped when all items fit on a single page.
-    """
-    total = len(items)
-    total_pages = math.ceil(total / page_size) if total > 0 else 1
-    page = 0
-
-    while True:
-        start = page * page_size
-        render_fn(items[start : start + page_size], page, total_pages, total)
-
-        if total_pages <= 1:
-            break
-
-        choices = []
-        if page < total_pages - 1:
-            choices.append(questionary.Choice("Next →", value="next"))
-        if page > 0:
-            choices.append(questionary.Choice("← Previous", value="prev"))
-        choices.append(questionary.Choice("Exit", value="quit"))
-
-        action = questionary.select("", choices=choices).ask()
-        if action is None or action == "quit":
-            break
-        elif action == "next":
-            page += 1
-        elif action == "prev":
-            page -= 1
-
-
-def _truncate(text: str, max_len: int = 60) -> str:
-    return text if len(text) <= max_len else text[: max_len - 1] + "…"
-
-
-def _read_tools_file(path: Path) -> list[str]:
-    """Read tool specs from a file, one per line.
-
-    Strips inline comments (anything after ``#``), blank lines, and
-    leading/trailing whitespace. Raises FileNotFoundError if path is missing.
-    """
-    tools = []
-    for raw in path.read_text(encoding="utf-8").splitlines():
-        line = raw.split("#", 1)[0].strip()
-        if line:
-            tools.append(line)
-    return tools
-
-
-def _render_tool_table(
-    results: list[tuple[str, str]],
-    title: str,
-    border_style: str = "primary",
-) -> None:
-    """Render a Rich table of (name, description) tool pairs."""
-    table = Table(
-        title=title,
-        box=ROUNDED,
-        border_style=border_style,
-        header_style="table.header",
-        show_lines=False,
-    )
-    table.add_column("Tool", style="tool", no_wrap=True)
-    table.add_column("Description", style="muted")
-    for name, desc in results:
-        table.add_row(name, _truncate(desc))
-    console.print(table)
-
-
-def _print_find_hint(source_note: str | None = None) -> None:
-    """Print the 'use shelley-bio find' footer line."""
-    suffix = f" · Source: {source_note}" if source_note else ""
-    console.print(
-        f"[muted]For more information about a specific tool, use "
-        f"[command]shelley-bio find <name>[/command]{suffix}[/muted]"
-    )
-
-
-def _render_search_page(
-    results: list[tuple[str, str]],
-    page: int,
-    total_pages: int,
-    total: int,
-    query: str,
-    cvmfs_filtered: bool = False,
-) -> None:
-    """Render one page of search results."""
-    count = total
-    suffix = "es" if count != 1 else ""
-    page_info = f" — page {page + 1} of {total_pages}" if total_pages > 1 else ""
-    title = f"[header]Results for '[tool]{query}[/tool]' ({count} match{suffix}){page_info}[/header]"
-    source_note = "RSEC bio.tools (CVMFS-available tools)" if cvmfs_filtered else "RSEC bio.tools"
-    _render_tool_table(results, title)
-    _print_find_hint(source_note=source_note)
-
-
-def search_tools(query: str) -> None:
-    """Search the RSEC corpus directly (no MCP server needed)."""
-    with ShelleyStyle.create_status(f"Searching for: {query}") as status:
-        try:
-            source = RsecSource().load()
-        except FileNotFoundError:
-            console.print(ShelleyStyle.create_error_panel(
-                "Corpus Not Found",
-                "rsec_meta.json.gz is missing.",
-                "Run: shelley-bio-build-rsec",
-            ))
-            return
-
-        cvmfs_ids = _load_cvmfs_tool_ids()
-        if cvmfs_ids is not None:
-            source.entries = [
-                e for e in source.entries
-                if e.get("id", "").lower().replace("-", "_") in cvmfs_ids
-            ]
-
-        names = source.search(query)
-
-    if not names:
-        console.print(ShelleyStyle.create_error_panel(
-            "No Results",
-            f"No tools matched '{query}'.",
-            "Try broader terms — e.g. 'alignment' instead of 'short-read alignment'",
-        ))
-        return
-
-    desc_for = {
-        str(e.get("name") or e.get("id") or ""): str(e.get("description") or "")
-        for e in source.entries
-    }
-    results = [(name, desc_for.get(name, "")) for name in names]
-
-    def render_page(page_items, page, total_pages, total):
-        _render_search_page(page_items, page, total_pages, total, query,
-                            cvmfs_filtered=(cvmfs_ids is not None))
-
-    _paginate(results, render_page)
-
-
-def _render_versions_page(
-    entries: list[dict],
-    page: int,
-    total_pages: int,
-    total: int,
-    tool_name: str,
-) -> None:
-    """Render one page of versions with buildable status."""
-    page_info = f" — page {page + 1} of {total_pages}" if total_pages > 1 else ""
-    table = Table(
-        title=f"[header]Available Versions for [tool]{tool_name}[/tool] ({total} total){page_info}[/header]",
-        box=ROUNDED,
-        border_style="primary",
-        header_style="table.header",
-        show_lines=False,
-    )
-    table.add_column("Version", style="version", no_wrap=True)
-    table.add_column("Buildable", no_wrap=True)
-    for entry in entries:
-        buildable_str = "[success]✓[/success]" if entry["buildable"] else "[muted]✗[/muted]"
-        table.add_row(entry["version"], buildable_str)
-    console.print(table)
-
-
-def versions_sync(tool_name: str) -> None:
-    """List all container versions for a tool from the CVMFS cache (no MCP server needed)."""
-    with ShelleyStyle.create_status(f"Loading versions for: {tool_name}") as status:
-        pairs = _load_versions_from_cache(tool_name)
-
-    if pairs is None:
-        console.print(ShelleyStyle.create_error_panel(
-            "Cache Not Found",
-            "galaxy_singularity_cache.json.gz is missing.",
-            "Run: shelley-bio-build-cache",
-        ))
-        return
-
-    if not pairs:
-        console.print(ShelleyStyle.create_error_panel(
-            "No Versions Found",
-            f"No containers found for '{tool_name}'.",
-            "Check the tool name spelling or try: shelley-bio find " + tool_name,
-        ))
-        return
-
-    entries = _compute_version_entries(tool_name, pairs)
-
-    def render_page(page_items, page, total_pages, total):
-        _render_versions_page(page_items, page, total_pages, total, tool_name)
-
-    _paginate(entries, render_page)
-
-    if any(not e["buildable"] for e in entries):
-        console.print(
-            "[muted]Buildable ✗: Versions not in the shpc registry may still be built, "
-            "but can take a few minutes longer. This is suited for users who need "
-            "a specific older version for reproducibility.[/muted]"
-        )
-    console.print(Panel(
-        f"To install the latest version of {tool_name}, run:\n\n"
-        f"[command]shelley-bio build {tool_name}[/command]",
-        title="[header]Install[/header]",
-        box=ROUNDED,
-        border_style="info",
-        padding=(0, 2),
-    ))
-
-
-async def query_tool(session: ClientSession, tool_name: str):
-    """Query for a specific tool."""
-    with ShelleyStyle.create_status(f"Searching for tool: {tool_name}") as status:
-        result = await session.call_tool("find_tool", {"tool_name": tool_name})
-
-    for content in result.content:
-        if hasattr(content, 'text'):
-            try:
-                payload = json.loads(content.text)
-                _render_find_tool(payload)
-            except json.JSONDecodeError:
-                console.print(content.text)
-
-
-async def list_tools(session: ClientSession, limit: int = 50):
-    """List available tools."""
-    with ShelleyStyle.create_status(f"Loading tools (limit: {limit})") as status:
-        result = await session.call_tool(
-            "list_available_tools",
-            {"limit": limit}
-        )
-    
-    for content in result.content:
-        if hasattr(content, 'text'):
-            console.print(content.text)
-
-
-def build_module(tool_spec: str) -> bool:
-    """Build an Lmod module for a tool from CVMFS.
-    
-    Returns:
-        bool: True if build was successful, False otherwise
-    """
-    # TODO: Why is this in cli.py?
-    # Check if we need sudo (by testing write access to module directory)
-    module_dir = Path("/apps/Modules/modulefiles")
-    needs_sudo = not os.access(module_dir, os.W_OK) if module_dir.exists() else True
-    
-    if needs_sudo:
-        # We need sudo - run the command with sudo automatically  
-        script_dir = Path(__file__).parent.parent.parent
-        shelley_bio_path = script_dir / "bin" / "shelley-bio"
-        
-        cmd = [
-            "sudo", "-E", "env", f"PATH={os.environ['PATH']}", 
-            str(shelley_bio_path), "build", tool_spec
-        ]
-        
-        try:
-            print_info(f"Running with elevated privileges: build {tool_spec}")
-            result = subprocess.run(cmd, check=False)
-            if result.returncode != 0:
-                print_error(f"Build failed with exit code {result.returncode}")
-                return False
-            return True
-        except subprocess.CalledProcessError as e:
-            print_error(f"Build command failed: {e}")
-            return False
-        except KeyboardInterrupt:
-            print_warning("Build cancelled by user")
-            return False
-    
-    # Original build logic for when we already have permissions
-    builder = CVMFSModuleBuilder()
-        
-    try:
-        # Get available versions first for display
-        if "/" in tool_spec:
-            tool_name, requested_version = tool_spec.split("/", 1)
-        elif ":" in tool_spec:
-            tool_name, requested_version = tool_spec.split(":", 1)
-        else:
-            tool_name, requested_version = tool_spec, None
-        
-        # Resolve version (may prompt interactively — must run outside any Live/spinner context)
-        final_tool, final_version = builder.search_tool_version(tool_name, requested_version)
-
-        # Install via shpc (creates wrapper scripts + .lua module file)
-        with ShelleyStyle.create_status(f"Building module for {tool_spec}") as status:
-            module_file = builder.shpc_install(final_tool, final_version)
-            available_versions = builder.list_versions(tool_name)
-        
-        # Display results
-        if requested_version is None and len(available_versions) > 1:
-            info_panel = ShelleyStyle.create_build_info(
-                final_tool, final_version, available_versions
-            )
-            console.print(info_panel)
-            print_rule()
-        
-        success_panel = ShelleyStyle.create_build_success(
-            final_tool, final_version, module_file
-        )
-        console.print(success_panel)
-        return True
-        
-    except Exception as e:
-        # TODO: This doesn't capture all the errors accurately, e.g. errors in CVMFSModuleBuilder methods
-        # TODO: Write unit test for e.g. build samtools:1.10293891 not found
-        # TODO: Also 
-        # Default message 
-        title="Build Failed"
-        msg=str(e)
-        suggestion="Check that CVMFS is mounted and the tool exists"
-
-        # Suppress generic suggestion for well-formed "not found" errors
-        if re.search(r"^Version .* not found for", msg):
-            suggestion=""
-
-        error_panel = ShelleyStyle.create_error_panel(
-            title=title,
-            message=msg,
-            suggestion=suggestion
-        )
-        console.print(error_panel)
-        return False
-
-def list_cvmfs_versions(tool_name: str) -> None:
-    """List available versions of a tool in CVMFS."""
-    builder = CVMFSModuleBuilder()
-    
-    try:
-        with ShelleyStyle.create_status(f"Scanning CVMFS for {tool_name} versions") as status:
-            version_path_pairs = builder.list_versions_with_paths(tool_name)
-            
-        if not version_path_pairs:
-            error_panel = ShelleyStyle.create_error_panel(
-                "No Versions Found",
-                f"No versions of '{tool_name}' found in CVMFS",
-                "Check the tool name spelling or try a different tool"
-            )
-            console.print(error_panel)
-        else:
-            versions_table = ShelleyStyle.create_versions_with_paths_table(tool_name, version_path_pairs)
-            console.print(versions_table)
-            
-    except Exception as e:
-        error_panel = ShelleyStyle.create_error_panel(
-            "CVMFS Access Error",
-            str(e),
-            "Ensure CVMFS is mounted at /cvmfs/singularity.galaxyproject.org/"
-        )
-        console.print(error_panel)
-
-
-def find_tool_sync(tool_name: str) -> None:
-    """Find a tool by name using RSEC + CVMFS cache directly (no MCP server needed)."""
-    clean_name = re.sub(r"[:/].*$", "", tool_name).strip()
-    query_lower = clean_name.lower()
-
-    with ShelleyStyle.create_status(f"Searching for tool: {clean_name}") as _status:
-        source = RsecSource().load()
-        pairs = _load_versions_from_cache(clean_name) or []
-
-    # Exact match by id or name (case-insensitive)
-    meta = None
-    for entry in source.entries:
-        if (entry.get("id", "").lower() == query_lower or
-                entry.get("name", "").lower() == query_lower):
-            meta = entry
-            break
-
-    # If query didn't match the canonical id, retry versions lookup with meta id
-    if meta and not pairs:
-        meta_id = meta.get("id", "")
-        if meta_id.lower() != query_lower:
-            pairs = _load_versions_from_cache(meta_id) or []
-
-    found = meta is not None or bool(pairs)
-
-    suggestions: list[tuple[str, str]] = []
-    if not found:
-        entry_map = {
-            e["id"].lower(): (e.get("name") or e["id"], e.get("description") or "")
-            for e in source.entries if e.get("id")
-        }
-        suggestions = [entry_map[m] for m in get_close_matches(query_lower, entry_map.keys(), n=8, cutoff=0.6)]
-
-    containers_payload = None
-    if pairs:
-        tool_id = meta.get("id", clean_name) if meta else clean_name
-        unique_versions = _compute_version_entries(tool_id, pairs)
-
-        containers_payload = {
-            "available": True,
-            "recent_versions": unique_versions[:5],
-            "total_versions": len(unique_versions),
-            "install_command": f"shelley-bio build {tool_id}",
-        }
-
-    tool_payload = None
-    if meta:
-        tool_payload = {
-            "id": meta.get("id", clean_name),
-            "name": meta.get("name", clean_name),
-            "description": meta.get("description") or "",
-            "homepage": meta.get("homepage") or "",
-            "operations": source._flatten_edam(meta.get("edam-operations")),
-            "inputs": source._flatten_edam(meta.get("edam-inputs")),
-            "outputs": source._flatten_edam(meta.get("edam-outputs")),
-        }
-
-    _render_find_tool({
-        "query": clean_name,
-        "found": found,
-        "suggestions": suggestions,
-        "tool": tool_payload,
-        "containers": containers_payload,
-    })
-
-
-async def interactive_mode(session: ClientSession):
-    """Run in interactive mode."""
-    console.clear()
-    print_banner()
-    
-    # Create help table
-    commands = [
-        {"command": "find <tool>", "description": "Find information about a specific tool", "example": "find fastqc"},
-        {"command": "search <description>", "description": "Search for tools by function", "example": "search quality control"},
-        {"command": "versions <tool>", "description": "Get available container versions", "example": "versions samtools"},
-        {"command": r"build <tool\[/ver]>", "description": "Build Lmod module for tool", "example": "build samtools/1.21"},
-        {"command": "help", "description": "Show detailed help", "example": "help"},
-        {"command": "exit", "description": "Exit interactive mode", "example": "exit"}
-    ]
-    
-    help_table = ShelleyStyle.create_help_table(commands)
-    console.print(help_table)
-    print_rule()
-    
-    while True:
-        try:
-            user_input = console.input("\n[prompt]shelley-bio>[/prompt] ").strip()
-            if not user_input:
-                continue
-                
-            parts = user_input.split()
-            command = parts[0].lower()
-            
-            if command in ["exit", "quit", "q"]:
-                print_success("🐢 Goodbye!")
-                break
-                
-            elif command == "help":
-                console.print(help_table)
-                console.print("\n")
-                console.print(ShelleyStyle.format_command_examples())
-                
-            elif command == "find" and len(parts) > 1:
-                loop = asyncio.get_running_loop()
-                await loop.run_in_executor(None, find_tool_sync, parts[1])
-                
-            elif command == "find":
-                print_warning("Missing tool name")
-                print_info("Usage: [command]find <tool_name>[/command]")
-                print_info("Example: [command]find fastqc[/command]")
-                
-            elif command == "search" and len(parts) > 1:
-                description = " ".join(parts[1:])
-                loop = asyncio.get_running_loop()
-                await loop.run_in_executor(None, search_tools, description)
-
-            elif command == "search":
-                print_warning("Missing search terms")
-                print_info("Usage: [command]search <description>[/command]")
-                print_info("Example: [command]search quality control[/command]")
-
-            elif command == "versions" and len(parts) > 1:
-                loop = asyncio.get_running_loop()
-                await loop.run_in_executor(None, versions_sync, parts[1])
-
-            elif command == "versions":
-                print_warning("Missing tool name")
-                print_info("Usage: [command]versions <tool_name>[/command]")
-                print_info("Example: [command]versions samtools[/command]")
-                
-            elif command == "build" and len(parts) > 1:
-                if build_module(parts[1]):
-                    print_success("Module built successfully! Exiting interactive mode.")
-                    break
-                    
-            elif command == "build":
-                print_warning("Missing tool specification")
-                print_info("Usage: [command]build <tool_name>[/version][/command]")
-                print_info("Examples:")
-                print_info("  [command]build fastqc[/command]")
-                print_info("  [command]build samtools/1.21[/command]")
-                
-            else:
-                print_warning(f"Unknown command or missing arguments: '{command}'")
-                print_info("Type [command]help[/command] for available commands")
-        
-        except KeyboardInterrupt:
-            print_info("\nExiting interactive mode...")
-            break
-        except EOFError:
-            print_info("\nExiting interactive mode...")
-            break
-        except Exception as e:
-            print_error(f"Error: {e}")
-
-
-async def _async_main() -> None:
-    """Handle commands that require the MCP server."""
-    # TODO: Contains too much duplicated code to main(). I don't think this is
-    # needed anymore as MCP is not used.
-    command = sys.argv[1].lower()
-
-    server_script = Path(__file__).parent.parent / "server" / "mcp_server.py"
-
-    if not server_script.exists():
-        error_panel = ShelleyStyle.create_error_panel(
-            "Server Configuration Error",
-            f"Server script not found at {server_script}",
-            "Check Shelley Bio installation"
-        )
-        console.print(error_panel)
-        sys.exit(1)
-    
-    # Server parameters
-    server_params = StdioServerParameters(
-        command="python3",
-        args=[str(server_script)],
-        env=None
-    )
-    
-    # Connect to server
-    try:
-        async with stdio_client(server_params) as (read, write):
-            async with ClientSession(read, write) as session:
-                # Initialize
-                await session.initialize()
-                
-                if command == "interactive":
-                    await interactive_mode(session)
-                
-                else:
-                    error_panel = ShelleyStyle.create_error_panel(
-                        "Unknown Command",
-                        f"Unknown command: {command}",
-                        "Use 'shelley-bio' with no arguments to see available commands"
-                    )
-                    console.print(error_panel)
-                    sys.exit(1)
-                    
-    except Exception as e:
-        error_panel = ShelleyStyle.create_error_panel(
-            "Connection Error",
-            f"Failed to run shelly-bio: {e}",
-            "Check that all dependencies are installed"
-        )
-        console.print(error_panel)
-        sys.exit(1)
 
 
 def _print_usage() -> None:
@@ -771,7 +29,6 @@ def _print_usage() -> None:
         {"command": "search <description>", "description": "Search for tools by function", "example": "shelley-bio search 'quality control'"},
         {"command": "versions <tool_name>", "description": "Get available container versions", "example": "shelley-bio versions samtools"},
         {"command": r"build <tool\[/version\]>", "description": "Build Lmod module for tool", "example": "shelley-bio build samtools/1.21"},
-        {"command": "interactive", "description": "Start interactive mode", "example": "shelley-bio interactive"},
         {"command": "help", "description": "Show this help message", "example": "shelley-bio help"},
     ]
 
@@ -779,13 +36,11 @@ def _print_usage() -> None:
     console.print(usage_table)
     console.print("\n")
     console.print(ShelleyStyle.format_command_examples())
-
     print_rule()
-    print_info("For interactive mode with guided commands: [command]shelley-bio interactive[/command]")
 
 
 def main() -> None:
-    """CLI entry point. Sync commands are handled here before entering the event loop."""
+    """CLI entry point."""
     if len(sys.argv) < 2:
         _print_usage()
         sys.exit(1)
@@ -800,7 +55,7 @@ def main() -> None:
         arg = sys.argv[2]
         p = Path(arg)
         if p.is_file():
-            tools = _read_tools_file(p)
+            tools = read_tools_file(p)
             if not tools:
                 print_warning(f"No tool specs found in '{arg}' (file is empty or all comments)")
                 sys.exit(1)
@@ -834,7 +89,12 @@ def main() -> None:
             print_info("Example: [command]shelley-bio versions samtools[/command]")
         sys.exit(0)
 
-    asyncio.run(_async_main()) #TODO: Not needed anymore without MCP?
+    console.print(ShelleyStyle.create_error_panel(
+        "Unknown Command",
+        f"Unknown command: '{command}'",
+        "Run shelley-bio --help for usage",
+    ))
+    sys.exit(1)
 
 
 if __name__ == "__main__":
