--- shelley_bio/server/__init__.py
@@ -1,7 +1,0 @@
-"""
-MCP Server module for bioinformatics tool discovery.
-"""
-
-from .mcp_server import main as server_main
-
-__all__ = ["server_main"]
